+++ worker.py
@@ -1,66 +1,28 @@
 #!/root/rabbitmq-scripts/env_pika/bin/python3
-import pika
+#import pika
 from PikaConn import PikaConn
 import sys, os, time
 import click
 
-@click.command()
-@click.option('--queue', '-q', default='queue name', show_default=True)
-@click.option('--exchange', '-e', default='', show_default=True)
-@click.option('--type', '-t', default='', show_default=True)
-@click.option('--exclusive', '-x', default=True, show_default=True)
-@click.option('--durable', '-d', default=False, show_default=True)
-def createConnection(queue, exchange, type, exclusive, durable):
-	exchangeTuple=(exchange, type)
-	conn = PikaConn(queueName = queue, 
-					exchange = exchangeTuple, 
-					durable = durable,
-					exclusive = exclusive,
-					sending = False)
-	return conn
+import pika
 
-def subscribe(pc):
-    connection = pika.BlockingConnection(pc.params)
-	channel = connection.channel()
-	if conn.isExchange:
-		channel.exchange_declare(exchange = conn.exchangeName, exchange_type= conn.exchangeType)
-		result = channel.queue_declare("", exclusive=conn.exclusive)
-		tempQueueName = result.method.queue
-		channel.queue_bind(exchange=conn.exchangeName, queue=tempQueueName)
-		print(' [*] Waiting for logs. To exit press CTRL+C')
+connection = pika.BlockingConnection(
+    pika.ConnectionParameters(host='localhost'))
+channel = connection.channel()
 
-		def callback(ch, method, properties, body):
-			print(f" [x] {body}")
+channel.exchange_declare(exchange='logs', exchange_type='fanout')
 
-		channel.basic_consume(
-			queue=tempQueueName, on_message_callback=callback, auto_ack=True)
+result = channel.queue_declare(queue='', exclusive=True)
+queue_name = result.method.queue
 
-	else:
-		channel.queue_declare(queue=conn.queueName, durable=conn.durable)
-		print(' [*] Waiting for logs. To exit press CTRL+C')
+channel.queue_bind(exchange='logs', queue=queue_name)
 
-		def callback(ch, method, properties, body):
-			print(f" [x] Received {body.decode()}")
-			time.sleep(body.count(b'.'))
-			print(" [x] Done")
-			ch.basic_ack(delivery_tag=method.delivery_tag)
+print(' [*] Waiting for logs. To exit press CTRL+C')
 
-		channel.basic_qos(prefetch_count=1)
-		channel.basic_consume(queue=conn.getQueueName(), on_message_callback=callback)
-  
-	return channel
+def callback(ch, method, properties, body):
+    print(f" [x] {body}")
 
-try:
-	conn = createConnection()
-	print(conn.queueName)
-	channel = subscribe(conn)
-	print("subscribed")
-	channel.start_consuming()
-except KeyboardInterrupt:
-	print('user interrupted')
-	try:
-		sys.exit(0)
-	except SystemExit:
-		os._exit(0)
-	finally:
-		conn.close()
+channel.basic_consume(
+    queue=queue_name, on_message_callback=callback, auto_ack=True)
+
+channel.start_consuming()
